Clean up of GUI implementation
